Move EffectType functions
